requirements.txt
msgpack>=1.1.0
ruff>=0.11.0
PyYAML>=6.0.2
